[Gold V] Title: Z, Time: 0 ms, Memory: 2020 KB -BaekjoonHub

+++ 백준/Gold/1074. Z/Z.cc
@@ -2,14 +2,16 @@
 using namespace std;
 
 int main() {
+
+
     int N;
-    long long r, c;
+    int r, c;
     cin >> N >> r >> c;
 
-    long long ans = 0;
+    int ans = 0;
     for (int i = 0; i < N; i++) {
-        long long rb = (r >> i) & 1LL;
-        long long cb = (c >> i) & 1LL;
+        int rb = (r >> i) & 1LL;
+        int cb = (c >> i) & 1LL;
         ans |= (rb << (2 * i + 1));
         ans |= (cb << (2 * i));
     }
